+++ artemis/src/test/kotlin/ktx/artemis/TestSystem.kt
@@ -1,0 +1,7 @@
+package ktx.artemis
+
+import com.artemis.BaseSystem
+
+class TestSystem : BaseSystem() {
+  override fun processSystem() = Unit
+}